Lift the different queues into their own file
