+++ calyx-py/calyx/pifo_oracle.py
@@ -1,91 +1,7 @@
+import queues
 import fifo_oracle
 
-from dataclasses import dataclass
-from typing import List, Tuple
-
 ANS_MEM_LEN = 10
-
-
-@dataclass
-class Pifo:
-    """A PIFO data structure.
-    Supports the operations `push`, `pop`, and `peek`.
-
-    We do this by maintaining two queues that are given to us at initialization.
-    We toggle between these queues when popping/peeking.
-    We have a variable called `hot` that says which queue is to be popped/peeked next.
-    `hot` starts at 1.
-
-    We maintain a variable called `pifo_len`: the sum of the lengths of the two queues.
-
-    When asked to pop:
-    - If `pifo_len` is 0, we raise an error.
-    - Else, if `hot` is 1, we try to pop from queue_1.
-      + If it succeeds, we flip `hot` to 2 and return the value we got.
-      + If it fails, we pop from queue_2 and return the value we got.
-        We leave `hot` as it was.
-    - If `hot` is 2, we proceed symmetrically.
-    - We decrement `pifo_len` by 1.
-
-    When asked to peek:
-    We do the same thing as above, except:
-    - We peek instead of popping.
-    - We don't flip `hot`.
-
-    When asked to push:
-    - If the value to be pushed is less than 200, we push it into queue_1.
-    - Else, we push it into queue_2.
-    - We increment `pifo_len` by 1.
-    """
-
-    def __init__(self, queue_1, queue_2):
-        self.data = (queue_1, queue_2)
-        self.hot = 1
-        self.pifo_len = 0
-
-    def push(self, val: int):
-        """Pushes `val` to the PIFO."""
-        if val < 200:
-            self.data[0].push(val)
-        else:
-            self.data[1].push(val)
-        self.pifo_len += 1
-
-    def pop(self) -> int:
-        """Pops the PIFO."""
-        if self.pifo_len == 0:
-            raise IndexError("Cannot pop from empty PIFO.")
-        if self.hot == 1:
-            try:
-                self.pifo_len -= 1
-                self.hot = 2
-                return self.data[0].pop()
-            except IndexError:
-                self.pifo_len -= 1
-                return self.data[1].pop()
-        else:
-            try:
-                self.pifo_len -= 1
-                self.hot = 1
-                return self.data[1].pop()
-            except IndexError:
-                self.pifo_len -= 1
-                return self.data[0].pop()
-
-    def peek(self) -> int:
-        """Peeks into the PIFO."""
-        if self.pifo_len == 0:
-            raise IndexError("Cannot peek into empty PIFO.")
-        if self.hot == 1:
-            try:
-                return self.data[0].peek()
-            except IndexError:
-                return self.data[1].peek()
-        else:
-            try:
-                return self.data[1].peek()
-            except IndexError:
-                return self.data[0].peek()
 
 
 def operate_pifo(commands, values):
@@ -94,7 +10,7 @@ def operate_pifo(commands, values):
     In the end, we return the answer memory.
     """
 
-    pifo = Pifo(fifo_oracle.Fifo([]), fifo_oracle.Fifo([]))
+    pifo = queues.Pifo(queues.Fifo([]), queues.Fifo([]))
     # Our PIFO is simple: it just orchestrates two FIFOs.
 
     ans = []
