+++ calyx-py/test/correctness/fifo.data
@@ -25,21 +25,21 @@
   },
   "values": {
     "data": [
-      47,
-      60,
-      31,
-      48,
-      69,
-      13,
-      73,
-      31,
-      1,
-      93,
-      27,
+      190,
+      240,
+      126,
+      194,
+      278,
       52,
-      35,
-      23,
-      98
+      293,
+      127,
+      6,
+      374,
+      110,
+      208,
+      143,
+      93,
+      392
     ],
     "format": {
       "is_signed": false,
