+++ calyx-py/test/correctness/fifo.expect
@@ -1,13 +1,13 @@
 {
   "ans_mem": [
-    47,
-    47,
-    47,
-    31,
-    31,
-    69,
-    13,
-    73,
+    190,
+    190,
+    190,
+    126,
+    126,
+    278,
+    52,
+    293,
     0,
     0
   ],
@@ -29,20 +29,20 @@
     0
   ],
   "values": [
-    47,
-    60,
-    31,
-    48,
-    69,
-    13,
-    73,
-    31,
-    1,
-    93,
-    27,
+    190,
+    240,
+    126,
+    194,
+    278,
     52,
-    35,
-    23,
-    98
+    293,
+    127,
+    6,
+    374,
+    110,
+    208,
+    143,
+    93,
+    392
   ]
 }
